WIP debugging tiling for charts with non-zero skew.
CoordsPolynomialTranslator.java
Changed CoordsPolynomialTranslator to accept an AffineTransform in the constructors.

The coordinates polynomial translators are written in x,y and lat,lon referencing the original chart. For charts with non-zero skew, this presents a problem.

The affine transform is the one that rotates the chart to North up and translates it so that it fits into a box starting at the upper-left corner of the screen .

Modified  public Point getAbsolutePointFromPosition(double lon, double lat) so that it returns the point in the North-up transformed chart.

Modified Position getAbsolutePositionFromPoint(Point q) so that it uses the inverse transform and takes a point in the transformed Nurth-up chart, transforms this point to the corresponding point in the original chart and then calculats the position from the polynomial.

Added logging support and logging statements that printout the transformed and untransformed points.

Position.java
Added a toString method.

MapReference.java
Added a toString() method to printout the Point and Position.

KapProcessor.java
Changed the log method so that it adds the current method to the log message.

Changed public void createTilePyramid(File kapFile, File pyramidPath) to generate the AffineTransform for non-zero skew charts. Reads and saves (as *.png) the chart files.

KAPParser.java
Added the AffineTransform as a private field.
Many changes in loadHeader to add logging statements, to get the BufferedImage from the file and to determine the proper AffineTransform.

Added explicit ligging of the displayLimits values.

Added addional comments and logging statements in public Sector extractVisibleSectorFromBounds(Vector<Position> bounds).