Chart converstion now working for most skewed charts but not all.
CoordsPolynomTransform
Redid the debugging statements.

KAPParser.java
Added a build of the CoordsPolynomGTransform when endOfSection is true
Added logger.isDebugEnabled blocks to generate and save the initial, rotated and finally rotated and translated png images.
If the program is run with debug level set to info, it will run faster and generate fewer files.
Corrected the calculation of the translation following rotation.
Added a rebuild of the CoordsPolynomTransform at the end so that it has the AffineTransform corresponding to the rotation and translations.
KAPParser now holds the rotated, translate chart in the BufferedImage image field.
Rewrote the section that calculates the upperLeft and bottomRight bounds of the rotated-translated chart.

KapProcessor.java
Corrected the log method to select the second element in the stack to get the currently executing method. It was usint the 3rd element which was incorrect.