Chart conversion is now working. Checked on 250 files from NOAA CGD07 and the suite of NC RNCs from NOAA.
Stil getting a Raster error but the offending chart converts properly. For the moment, I will ignore the error.

KAPParser.java
Added a logging statement to printout the Position vlaues read in teh PLY section of loadHEader()
Changed the AffineTransform roatation to rotate around the center of the skewed chart. In this way, it cannot be rotated out of the visible screen.
Added the method getBounds(Point2D.Double []) to return the bounding Rectangle2D.Double.

KapParser.java
in createTilePyramid, removed any "&" in the title string. If and amperstand is present, you get SAXE parse errors.
In createTiles, if the image width or height it <= 0, return. Oterwise getImage will fail.